# src/web_app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sample API",
    "description": "A sample API that demonstrates how to use Flask-SwaggerUI.",
    "version": "1.0"
  },
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users/guests": {
      "get": {
        "summary": "get all guest in the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          }
        }
      }
    },
    "/users/guests/{guest_id}": {
      "get": {
        "summary": "get a guest in the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          },
          {
            "name": "guest_id",
            "in": "path",
            "description": "guest identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
                "$ref": "#/definitions/user"
            }
          },
          "404": {
            "description": "guest not found.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "update a guest in the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          },
          {
            "name": "guest_id",
            "in": "path",
            "description": "guest identifier",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "user data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_update"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
                "$ref": "#/definitions/user"
            }
          },
          "404": {
            "description": "guest not found.",
            "schema": {
                "$ref": "#/definitions/error"
            }
          }
        }
      },
      "delete": {
        "summary": "delete a guest from the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          },
          {
            "name": "guest_id",
            "in": "path",
            "description": "guest identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response.",
            "schema": {
              "type": "object",
              "properties": {
              }
            }
          },
          "404": {
            "description": "guest not found.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/users/employees": {
      "get": {
        "summary": "get all employees in the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          }
        }
      }
    },
    "/users/employees/{employee_id}": {
      "get": {
        "summary": "get a employee in the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          },
          {
            "name": "employee_id",
            "in": "path",
            "description": "employee identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
                "$ref": "#/definitions/user"
            }
          },
          "404": {
            "description": "employee not found.",
            "schema": {
                "$ref": "#/definitions/error"
            }
          }
        }
      },
      "patch": {
        "summary": "update a employee in the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          },
          {
            "name": "employee_id",
            "in": "path",
            "description": "employee identifier",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "user data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user_update"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "schema": {
                "$ref": "#/definitions/user"
            }
          },
          "404": {
            "description": "employee not found.",
            "schema": {
                "$ref": "#/definitions/error"
            }
          }
        }
      },
      "delete": {
        "summary": "delete a employee from the system.",
        "tags": ["users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "JWT access token",
            "required": true,
            "type": "string"
          },
          {
            "name": "employee_id",
            "in": "path",
            "description": "employee identifier",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response.",
            "schema": {
              "type": "object",
              "properties": {
              }
            }
          },
          "404": {
            "description": "employee not found.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
...
    "user": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "logged_in": {
          "type": "boolean"
        }
      }
    },
    "user_update": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
  }
}